refactor: 폴더와 파일 이름 정리
