+++ src/components/common/feature-list-application/ListApplication.module.scss
