+++ src/components/common/ui-card-notice/CardNotice.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import classNames from 'classnames/bind';
+import styles from './CardNotice.module.scss';
+
+const cx = classNames.bind(styles);
+
+export const CardNotice = () => <div className={cx('container')}></div>;
