+++ src/components/common/ui-detail-layout/DetailLayout.module.scss
@@ -1,4 +1,4 @@
-@import "@/src/styles/global.scss";
+@import '@/src/styles/global.scss';
 
 .container {
   border-radius: 12px;
