+++ src/components/common/ui-detail-layout/DetailLayout.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import classNames from 'classnames/bind';
+import styles from './DetailLayout.module.scss';
+
+const cx = classNames.bind(styles);
+
+export const DetailLayout = () => <div className={cx('container')}></div>;
