--- src/components/common/ui-notice-card/NoticeCard.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import classNames from 'classnames/bind';
-import styles from './NoticeCard.module.scss';
-
-const cx = classNames.bind(styles);
-
-export const NoticeCard = () => <div className={cx('container')}></div>;
