--- src/components/common/ui-profile-card/ProfileCard.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import classNames from 'classnames/bind';
-import styles from './ProfileCard.module.scss';
-
-const cx = classNames.bind(styles);
-
-export const ProfileCard = () => <div className={cx('container')}></div>;
