+++ src/components/common/ui-table/TableForOwner.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames/bind';
 import { items } from '@/src/types/types';
-import { TableButton } from '../ui-table-button/TableButton';
+import { StatusTag } from '../ui-status-tag/StatusTag';
 import styles from './Table.module.scss';
 
 const cx = classNames.bind(styles);
@@ -29,7 +29,7 @@ export const TableForOwner: TableType = ({ applications }) => (
           <td>{application.item.user.item.bio ?? '.'}</td>
           <td>{application.item.user.item.phone ?? '.'}</td>
           <td>
-            <TableButton status={application.item.status} />{' '}
+            <StatusTag status={application.item.status} />{' '}
             {/* TODO: status pending 일 때, 사장과 알바 각각 '거절하기'와 '승인하기' 버튼 & '취소하기' 버튼 */}
           </td>
         </tr>
