+++ src/layouts/list-layout/ListLayout.tsx
@@ -1,7 +1,7 @@
 import classNames from 'classnames/bind';
 import { Layout } from '../feature-layout/Layout';
 import styles from './ListLayout.module.scss';
-import { ApplicationList } from '@/src/components/common/feature-application-list/ApplicationList';
+import { ApplicationList } from '@/src/components/common/feature-list-application/ListApplication';
 
 const cx = classNames.bind(styles);
